% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelp.R
\name{kelp}
\alias{kelp}
\title{Create a kelp object for interacting with SeaweedFS.}
\description{
Create a kelp object for interacting with SeaweedFS.

Create a kelp object for interacting with SeaweedFS.
}
\details{
This is a slightly higher level abstraction than \code{\link[=seaweed_master]{seaweed_master()}} and
\code{\link[=seaweed_volume]{seaweed_volume()}}.
}
\keyword{export}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{master}}{A \link{seaweed_master} object for sending requests}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{kelp$new()}}
\item \href{#method-upload}{\code{kelp$upload()}}
\item \href{#method-download}{\code{kelp$download()}}
\item \href{#method-delete}{\code{kelp$delete()}}
\item \href{#method-delete_collection}{\code{kelp$delete_collection()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create client object for sending http requests to seaweed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kelp$new(seaweed_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seaweed_url}}{Root URL of Seaweed}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{kelp} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload"></a>}}
\if{latex}{\out{\hypertarget{method-upload}{}}}
\subsection{Method \code{upload()}}{
Upload file to SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kelp$upload(path, collection = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to file to be uploaded}

\item{\code{collection}}{Collection name, acts as a namespace for files.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The uploaded file ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
Download file from SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kelp$download(fid, path = tempfile(), collection = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fid}}{SeaweedFS file ID to read}

\item{\code{path}}{Path to download file to}

\item{\code{collection}}{Optional collection name this file belongs to.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The file contents
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete file from SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kelp$delete(fid, collection = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fid}}{SeaweedFS file ID to delete}

\item{\code{collection}}{Optional collection name this file belongs to.
This helps speedup lookup - only the single fid will be deleted.
See \href{#method-delete_collection}{
\code{kelp$delete_collection()}} to remove an entire
collection.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, called for side effects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_collection"></a>}}
\if{latex}{\out{\hypertarget{method-delete_collection}{}}}
\subsection{Method \code{delete_collection()}}{
Delete a collection of files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kelp$delete_collection(collection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection}}{Collection name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, called for side effects
}
}
}
