% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaweed_master.R
\name{seaweed_master}
\alias{seaweed_master}
\title{Seaweed Master API}
\description{
Create a \code{seaweed_master} object for interacting with SeaweedFS master API

See \url{https://github.com/chrislusf/seaweedfs/wiki/Master-Server-API} for
details
}
\keyword{export}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client}}{A \code{seaweed_client} object for sending requests}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{seaweed_master$new()}}
\item \href{#method-assign}{\code{seaweed_master$assign()}}
\item \href{#method-lookup}{\code{seaweed_master$lookup()}}
\item \href{#method-delete_collection}{\code{seaweed_master$delete_collection()}}
\item \href{#method-upload}{\code{seaweed_master$upload()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create client object for sending http requests to seaweed master
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seaweed_master$new(seaweed_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seaweed_url}}{Root URL of Seaweed master}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{seaweed_master} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assign"></a>}}
\if{latex}{\out{\hypertarget{method-assign}{}}}
\subsection{Method \code{assign()}}{
Assign a file key.

Increase a number in master server's memory and return fid, volume
server url and volume server public URL. Can be used to to upload
a file to \verb{<publicUrl>/<fid>}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seaweed_master$assign(collection = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection}}{Collection name, acts as a namespace for files.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
fid, volume server URL and volume server public URL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lookup"></a>}}
\if{latex}{\out{\hypertarget{method-lookup}{}}}
\subsection{Method \code{lookup()}}{
Lookup volume

Look up volumes by ID or by file ID to get volume URL. Volumes
might move so we need to lookup their URL to get up to date location.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seaweed_master$lookup(id, collection = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{volume ID or file ID to locate volume for}

\item{\code{collection}}{Optional collection name this volume belongs to.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of \code{\link[=seaweed_volume]{seaweed_volume()}} objects representing the volumes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_collection"></a>}}
\if{latex}{\out{\hypertarget{method-delete_collection}{}}}
\subsection{Method \code{delete_collection()}}{
Delete a collection of files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seaweed_master$delete_collection(collection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection}}{Collection name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, called for side effects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload"></a>}}
\if{latex}{\out{\hypertarget{method-upload}{}}}
\subsection{Method \code{upload()}}{
Upload file to SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seaweed_master$upload(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to file to be uploaded}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The uploaded location, name, URL and size
}
}
}
