% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaweed_volume.R
\name{seaweed_volume}
\alias{seaweed_volume}
\title{Seaweed Volume API}
\description{
Create a \code{seaweed_volume} object for interacting with a SeaweedFS volume

See \url{https://github.com/chrislusf/seaweedfs/wiki/Volume-Server-API} for
details
}
\keyword{export}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{seaweed_volume$new()}}
\item \href{#method-upload_file}{\code{seaweed_volume$upload_file()}}
\item \href{#method-upload_object}{\code{seaweed_volume$upload_object()}}
\item \href{#method-read}{\code{seaweed_volume$read()}}
\item \href{#method-download_file}{\code{seaweed_volume$download_file()}}
\item \href{#method-download_object}{\code{seaweed_volume$download_object()}}
\item \href{#method-delete}{\code{seaweed_volume$delete()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create client object for sending http requests to seaweed volume
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seaweed_volume$new(seaweed_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seaweed_url}}{Root URL of Seaweed volume}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{seaweed_volume} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload_file"></a>}}
\if{latex}{\out{\hypertarget{method-upload_file}{}}}
\subsection{Method \code{upload_file()}}{
Upload a file to SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seaweed_volume$upload_file(fid, path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fid}}{The file ID from SeaweedFS. See
\href{../../kelp/html/seaweed_master.html#method-assign}{
\code{seaweed_master$assign()}}
to get a file ID. Or use
\href{../../kelp/html/seaweed_master.html#method-upload_file}{
\code{seaweed_master$upload_file()}} to upload
directly.}

\item{\code{path}}{Path to file to upload}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Size of uploaded file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload_object"></a>}}
\if{latex}{\out{\hypertarget{method-upload_object}{}}}
\subsection{Method \code{upload_object()}}{
Upload an arbitrary R object to SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seaweed_volume$upload_object(fid, object)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fid}}{The file ID from SeaweedFS. See
\href{../../kelp/html/seaweed_master.html#method-assign}{
\code{seaweed_master$assign()}}
to get a file ID. Or use
\href{../../kelp/html/seaweed_master.html#method-upload_file}{
\code{seaweed_master$upload_file()}} to upload
directly.}

\item{\code{object}}{The object to upload}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Size of uploaded object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Read file from SeaweedFS into R
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seaweed_volume$read(fid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fid}}{SeaweedFS file ID to read}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The file contents
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_file"></a>}}
\if{latex}{\out{\hypertarget{method-download_file}{}}}
\subsection{Method \code{download_file()}}{
Download file from SeaweedFS to a local path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seaweed_volume$download_file(fid, path = tempfile())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fid}}{SeaweedFS file ID to download}

\item{\code{path}}{Local file path to save to}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The file path written to
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_object"></a>}}
\if{latex}{\out{\hypertarget{method-download_object}{}}}
\subsection{Method \code{download_object()}}{
Download and deserialize R object from SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seaweed_volume$download_object(fid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fid}}{SeaweedFS file ID to download}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The deserialized object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete a file from SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seaweed_volume$delete(fid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fid}}{SeaweedFS file ID to delete}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, called for side effects
}
}
}
