% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelp.R
\name{kelp}
\alias{kelp}
\title{Create a kelp object for interacting with SeaweedFS}
\description{
Create a kelp object for interacting with SeaweedFS

Create a kelp object for interacting with SeaweedFS
}
\keyword{export}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{seaweed_client}}{A \code{seaweed_client} object for sending requests}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{kelp$new()}}
\item \href{#method-upload}{\code{kelp$upload()}}
\item \href{#method-file_url}{\code{kelp$file_url()}}
\item \href{#method-read}{\code{kelp$read()}}
\item \href{#method-delete}{\code{kelp$delete()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create client object for sending http requests to seaweed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kelp$new(seaweed_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seaweed_url}}{Root URL of Seaweed}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{kelp} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload"></a>}}
\if{latex}{\out{\hypertarget{method-upload}{}}}
\subsection{Method \code{upload()}}{
Upload file to SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kelp$upload(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to file to be uploaded}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The uploaded location, name, URL and size
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-file_url"></a>}}
\if{latex}{\out{\hypertarget{method-file_url}{}}}
\subsection{Method \code{file_url()}}{
Get file URL from file ID.

This gets URL that can be used to read, update or delete a file. We
don't just use the URL returned on file creation as the volume may
have moved.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kelp$file_url(fid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fid}}{SeaweedFS file ID}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
URL which can be used to read, update or delete a file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Read file from SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kelp$read(fid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fid}}{SeaweedFS file ID to read}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The file contents
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete file from SeaweedFS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kelp$delete(fid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fid}}{SeaweedFS file ID to delete}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Response from SeaweedFS, the deleted size
}
}
}
